requirements.txt
# CodeSensei Requirements
# Updated for Python 3.13 compatibility

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.8.0  # Updated for Python 3.13 compatibility

# LLM Integration
groq==0.4.1

# Static Analysis Tools
pylint==3.0.3
bandit==1.7.5
radon==6.0.1

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3

# CLI Tools (NEW - Day 15-17)
click==8.1.7
rich==13.7.0

# Optional (for development)
httpx==0.25.2  # For testing FastAPI
requests==2.31.0  # For API testing

# Add to existing requirements.txt
prometheus-client==0.19.0